fix(ExecutionContext): changes signature to write 4 bytes instead of 8
Prior, we passed a 'size_t' to allocateVariableSizedData(), which we used to write to the arena (8 bytes). However, we expected to write 4 bytes. This could lead to a scenario in which we overwrote bytes of the next buffer, e.g., bytes of the control block.

+++ nes-runtime/src/ExecutionContext.cpp
@@ -107,7 +107,7 @@ nautilus::val<int8_t*> ArenaRef::allocateMemory(const nautilus::val<size_t>& siz
     return result;
 }
 
-VariableSizedData ArenaRef::allocateVariableSizedData(const nautilus::val<size_t>& sizeInBytes)
+VariableSizedData ArenaRef::allocateVariableSizedData(const nautilus::val<uint32_t>& sizeInBytes)
 {
     auto basePtr = allocateMemory(sizeInBytes + nautilus::val<size_t>(4));
     *(static_cast<nautilus::val<uint32_t*>>(basePtr)) = sizeInBytes;
